+++ src/dynamics/time_integration.jl
@@ -372,12 +372,12 @@ function time_stepping!(
     end
     
     # UNSCALE, CLOSE, FINISH
-    finish!(feedback)                       # finish the progress meter, do first for benchmark accuracy
+    finalize!(feedback)                     # finish the progress meter, do first for benchmark accuracy
     unscale!(progn)                         # undo radius-scaling for vor, div from the dynamical core
     unscale!(diagn)                         # undo radius-scaling for vor, div from the dynamical core
-    close(output)                           # close netCDF file
+    finalize!(output, progn, diagn, model)  # possibly post-process output, then close netCDF file
     write_restart_file(output, progn)       # as JLD2 
-    finish!(model.callbacks, progn, diagn, model)
+    finalize!(model.callbacks, progn, diagn, model)
 
     # return a UnicodePlot of surface vorticity
     surface_vorticity = diagn.grid.vor_grid[:, end]